Validate Detector Rest Handler
